Move version file to pakyow namespace
