+++ pakyow.gemspec
@@ -1,4 +1,4 @@
-require File.expand_path('../lib/version', __FILE__)
+require File.expand_path('../lib/pakyow/version', __FILE__)
 
 Gem::Specification.new do |spec|
   spec.name                   = 'pakyow'
